subscriptions deserve their own module now
