+++ presenter/src/recognizer.rs
@@ -1,27 +1,17 @@
 use crate::{GestureRecognizer, InputState};
-use emergent_ui::{MouseButton, WindowMessage};
+use emergent_ui::WindowMessage;
 pub use mover::MoverRecognizer;
-use std::collections::HashSet;
 
 pub mod pan;
 pub use pan::PanRecognizer;
+
+mod subscriptions;
+pub use subscriptions::*;
 
 pub mod tap;
 pub use tap::TapRecognizer;
 
 pub mod mover;
-
-pub type Subscriptions = HashSet<Subscription>;
-
-#[derive(Copy, Clone, PartialEq, Eq, Hash, Debug)]
-pub enum Subscription {
-    /// Subscribed to continuity events after a specific mouse button was pressed.
-    ///
-    /// These are automatically subscribed for every recognizer that receives button presses.
-    ButtonContinuity(MouseButton),
-    /// Subscribed to timer ticks once per frame.
-    Ticks,
-}
 
 // Below follows a rather convoluted way of transporting a gesture recognizer including its subscription
 // state through a `Box<Any>`.
